Change naming of component
