+++ src/components/conversation/conversation-send-message-box/index.js
@@ -1,0 +1,2 @@
+import ConversationSendMessageBox from './ConversationSendMessageBox';
+export default ConversationSendMessageBox;
