--- src/components/conversation/conversation-textarea/index.js
@@ -1,3 +1,0 @@
-import
-ConversationTextarea from './ConversationTextarea';
-export default ConversationTextarea;
